move batch to lib
