+++ lib/pubsub.ts
@@ -1,5 +1,6 @@
 import { v4 as uuidv4 } from "uuid";
 
+// TODO: Pubsub Generic Refactor
 let instance: PubSub | undefined;
 
 export class PubSub {
